Update README.md
Hibernation data written by macOS to the RTC is cleared by the BIOS and invalidates the hibernation file. This causes issues when rebooting the machine, as macOS cannot restore the previous session.

This issue is addressed by [HibernationFixup](https://github.com/acidanthera/HibernationFixup), which enables hibernation on systems with native NVRAM support.